LPD-59060 Merge FilterTranslator/SolrFilterTranslator into pojo
